Fixed problem with public_columns
Contest and Participant should not return token-fields to the user. Therefore they use scoping to exclude those field. Because scoping can only be used with collection of records (ActiveRecordRelation), finding a single record by token needs reading the id first with a condition on the token, and secondly reading of the record by id with scoping.