Started processing of update of Match.result
Added redundant attributes to the db for printed result, because there seems to be no way to define virtual attributes that are included into the activerecord_relation.
Those attributes are filled in a before_save callback.

BUT: The result that is sent in params ends up as being "unpermitted" and is not updated! There seem to be problems with string or array types.