Improved handling of Match.result
If a string is given, it is converted with JSON.parse, if an array is sent, it is not saved, because permit does not accept it!
Problems sometimes occured with whitespace in the string that contained non-breaking spaces. Since this field usually should not be keyed in but computed by the frontend, no correction were yet made.
Console log leves where adapted for dev, test and production (debug, warn, error).
Error messages in development were changed to :api instead of :html