Counts are computed after result update
The match stores a hash 'counts' of points, matches, sets and games won, tied or lost in its contesttypes_params
After saving the match, the ProcessManager sums up these counts for all matches of the contest and stores them in the Participants contesttype_params.
Those params can be used by the frontend to display a ranking list without having to compute all sums of its own. The counts will probably be extended with a 'rank' field to allow direct sorting.
Most of this processing has been refactored into the Result class to keep the ProcessManager clean of all the details.