Handle login with token similar to user login
Renamed email to username, which is more general (content can be both, is a frontend decision what to use).
Adaption of some tests and fixtures.

Login with token not with with a field in the link but in a second signin form. The request is handled similar to a user login, i.e. it gives back a csrf-token as session id.
Also returned is now some additional data, i.e. the type of signin (user/token) and the user record of the logged in user or some info to the used token (read/write, contest/participant, id).