Compute and save participant rank within the group
Computation of the rankvalue was taken over from the frontend, so the frontend does not have to compute the rank itself any more.
Ranks can be equal for several players and then have gaps, e.g. 1, 2, 2, 4.
Added contest 'Demoturnier Gruppen' with 7 (famous) participants to fixtures in order to have a test contest with 2 groups.